pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "obsidian-livesync-mcp"
version = "0.1.0"
description = "MCP server and CLI for Obsidian vault via CouchDB LiveSync — no app required"
requires-python = ">=3.10"
license = "MIT"
readme = "README.md"
authors = [
    { name = "Dan Staples" },
    { name = "Suhas Vemuri" },
]
keywords = ["obsidian", "mcp", "couchdb", "livesync", "cli", "vault"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries",
]
dependencies = [
    "mcp[cli]>=1.12.0",
    "httpx>=0.27.0",
    "pyyaml>=6.0",
    "pydantic>=2.0",
    "xxhash>=3.0",
]

[project.urls]
Homepage = "https://github.com/dismantl/obsidian-livesync-mcp"
Repository = "https://github.com/dismantl/obsidian-livesync-mcp"
Issues = "https://github.com/dismantl/obsidian-livesync-mcp/issues"

[project.scripts]
obsidian = "obsidian_livesync_mcp.cli:main"

[project.optional-dependencies]
oauth = [
    "pyjwt[crypto]>=2.10.1",
]
dev = [
    "ruff>=0.9.0",
    "pytest>=8.0",
    "pytest-asyncio>=0.24.0",
    "respx>=0.22.0",
    "pyjwt[crypto]>=2.10.1",
    "cryptography>=43.0",
]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "W", "I"]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.setuptools.packages.find]
where = ["src"]
